requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.59